feat : 지출 통계 API (3) 다른 유저 대비 소비율

+++ src/main/java/com/hyerijang/dailypay/statistics/dto/StatisticsDto.java
@@ -4,6 +4,6 @@ import com.hyerijang.dailypay.budget.domain.Category;
 import java.util.Map;
 
 public record StatisticsDto(Long totalExpenseComparison,
-                            Map<Category, Long> categoryExpenseComparison) {
+                            Map<Category, Double> categoryExpenseComparison) {
 
 }
